refactor: remove test-only exports and extract internal modules
- Remove __testing export bags from runner, cli, attach, runs-tui
- Privatize internal constants (baseBranchNames, limitsPollMs, noChangesReply)
- Privatize advisor bridge handlers (handleAdvise, handleFeedback)
- Extract internal modules: attach-runtime, runs-browser, limits-auth, coverage-core
- Delete duplicate tests from test/runner-extended.test.ts
- Refactor tests to use public APIs instead of internal exports
- Maintain 91.4% line coverage and 93.27% function coverage
