+++ src/attach.ts
@@ -1,19 +1,14 @@
-import { dirname, join } from "node:path"
+import { join } from "node:path"
 import { stdout } from "node:process"
 
-import { readRunMetadata, type PhaseMetadata, type RunMetadata } from "./metadata"
+import { LiveAttach, overallStatus, reconcileAdvisorJournal, replayHistory } from "./attach-runtime"
+import { readRunMetadata } from "./metadata"
 import { connectOpencode } from "./opencode"
 import { isServerLive } from "./runs"
-import { progressPhases, watchSession, type SessionWatcher } from "./runner"
+import { progressPhases } from "./runner"
 import { stepRunnerFor } from "./step-runners"
 import { createTuiProgress } from "./tui"
 import { runsRoot } from "./workspace"
-import { aggregateAdvisorEvents, readAdvisorEvents } from "./advisor-events"
-
-import type { OpencodeClient } from "@opencode-ai/sdk/v2"
-import type { ProgressPhaseSnapshot, ProgressUI } from "./progress"
-
-const pollMs = 1_000
 
 /**
  * Re-enters a run's convoy dashboard from `convoy runs`, without resuming it:
@@ -74,15 +69,12 @@ export async function openRunDashboard(runID: string): Promise<void> {
   tui.start(runID, targetDir, dir)
 
   if (!server) {
-    // Stopped run: paint history, then hand over to the finish screen.
     replayHistory(tui, metadata)
     await Promise.race([tui.runFinished?.({ status: overallStatus(metadata), runDir: dir }) ?? Promise.resolve(), detached])
     tui.stop()
     return
   }
 
-  // Live attach. Phases whose runner lacks live-attach support still update
-  // from metadata, but never open an OpenCode watcher.
   tui.serverReady(server.url)
   const phasesWithoutLiveAttach = new Set(phases.filter((phase) => !stepRunnerFor(phase.runner).capabilities.liveAttach).map((phase) => phase.name))
   const attach = new LiveAttach(connectOpencode(server.url), tui, targetDir, metaPath, phasesWithoutLiveAttach)
@@ -91,8 +83,6 @@ export async function openRunDashboard(runID: string): Promise<void> {
   await Promise.race([detached, attach.serverGone])
   await attach.stop()
 
-  // If the run ended while we watched (not a user detach), let the user keep
-  // browsing on the finish screen until they close it.
   if (!userDetached) {
     const latest = (await readRunMetadata(metaPath)) ?? metadata
     await reconcileAdvisorJournal(latest, dir)
@@ -100,155 +90,4 @@ export async function openRunDashboard(runID: string): Promise<void> {
     await Promise.race([tui.runFinished?.({ status: overallStatus(latest), runDir: dir }) ?? Promise.resolve(), detached])
   }
   tui.stop()
-}
-
-// Mirrors a live run's opencode activity into the dashboard: history from
-// metadata, plus a watchSession per running phase whose events stream in. All
-// read-only — it never drives the run, only observes it.
-export class LiveAttach {
-  readonly serverGone: Promise<void>
-  private resolveServerGone!: () => void
-  private readonly watchers = new Map<string, SessionWatcher>()
-  private readonly started = new Set<string>()
-  private readonly sessions = new Set<string>()
-  private readonly finalized = new Set<string>()
-  private poll?: ReturnType<typeof setInterval>
-  private stopped = false
-
-  constructor(
-    private readonly client: OpencodeClient,
-    private readonly tui: ProgressUI,
-    private readonly targetDir: string,
-    private readonly metaPath: string,
-    private readonly phasesWithoutLiveAttach: ReadonlySet<string> = new Set(),
-  ) {
-    this.serverGone = new Promise((resolve) => {
-      this.resolveServerGone = resolve
-    })
-  }
-
-  async start() {
-    await this.tick() // paint history before the caller starts waiting
-    this.poll = setInterval(() => void this.tick(), pollMs)
-    this.poll.unref?.()
-  }
-
-  private async tick() {
-    if (this.stopped) return
-    const metadata = await readRunMetadata(this.metaPath)
-    if (!metadata) return
-
-    // Metadata writes are debounced, while advisor events are appended
-    // immediately. Reconcile the journal on every poll so an attached viewer
-    // sees activity as it occurs rather than only after phase completion.
-    await reconcileAdvisorJournal(metadata, dirname(this.metaPath))
-
-    for (const [name, phase] of Object.entries(metadata.phases)) {
-      if (phase.sessionID && !this.sessions.has(name)) {
-        this.tui.phaseSession(name, phase.sessionID)
-        this.sessions.add(name)
-      }
-      if (phase.status === "running") {
-        if (!this.started.has(name)) {
-          this.started.add(name)
-          this.tui.phaseStarted(name)
-          if (phase.model) this.tui.phaseAttempt(name, { attempt: 1, model: phase.model })
-        }
-        for (const event of phase.advisorEvents ?? []) this.tui.phaseAdvisorEvent(name, event)
-        if (!this.phasesWithoutLiveAttach.has(name)) this.watch(name, phase.sessionID)
-      } else if (phase.status !== "pending" && !this.finalized.has(name)) {
-        // Reached a terminal state: stamp its real duration/cost and stop
-        // mirroring its (now finished) session.
-        this.finalized.add(name)
-        this.tui.phaseRestored(name, snapshotOf(phase, phase.status))
-        this.drop(name)
-      }
-    }
-
-    // The run cleared its server entry on shutdown: it's over.
-    if (!metadata.server && !this.stopped) this.resolveServerGone()
-  }
-
-  private watch(name: string, sessionID: string | undefined) {
-    if (!sessionID || this.watchers.has(name)) return
-    const watcher = watchSession(this.client, {
-      directory: this.targetDir,
-      phaseName: name,
-      sessionID,
-      progress: this.tui,
-      signal: new AbortController().signal, // stopped explicitly via watcher.stop()
-    })
-    this.watchers.set(name, watcher)
-    // When the session settles (the phase finished), stop mirroring it; the
-    // metadata poll finalizes its display.
-    watcher.result.then(
-      () => this.drop(name),
-      () => this.drop(name),
-    )
-  }
-
-  private drop(name: string) {
-    const watcher = this.watchers.get(name)
-    if (!watcher) return
-    this.watchers.delete(name)
-    void watcher.stop().catch(() => {})
-  }
-
-  async stop() {
-    if (this.stopped) return
-    this.stopped = true
-    if (this.poll) clearInterval(this.poll)
-    await Promise.all([...this.watchers.values()].map((watcher) => watcher.stop().catch(() => {})))
-    this.watchers.clear()
-  }
-}
-
-/** Merge the authoritative append-only journal over metadata's convenience projection. */
-async function reconcileAdvisorJournal(metadata: RunMetadata, runDir: string) {
-  const advisorEvents = await readAdvisorEvents(runDir)
-  for (const name of new Set(advisorEvents.map((event) => event.phase))) {
-    const events = advisorEvents.filter((event) => event.phase === name)
-    const phase = (metadata.phases[name] ??= { status: "pending" })
-    phase.advisorEvents = events
-    phase.advisor = aggregateAdvisorEvents(events)
-  }
-}
-
-// Replays every non-pending phase from metadata as a restored phase. A stale
-// "running" (a run interrupted mid-phase) reads as failed — it didn't finish.
-function replayHistory(tui: ProgressUI, metadata: RunMetadata) {
-  for (const [name, phase] of Object.entries(metadata.phases)) {
-    if (phase.sessionID) tui.phaseSession(name, phase.sessionID)
-    if (phase.status === "pending") continue
-    const status = phase.status === "running" ? "failed" : phase.status
-    tui.phaseRestored(name, snapshotOf(phase, status))
-  }
-}
-
-function snapshotOf(phase: PhaseMetadata, status: ProgressPhaseSnapshot["status"]): ProgressPhaseSnapshot {
-  return {
-    status,
-    sessionID: phase.sessionID,
-    durationMs: phase.durationMs,
-    cost: phase.cost,
-    tokens: phase.tokens,
-    model: phase.model,
-    advisor: phase.advisor,
-    advisorEvents: phase.advisorEvents,
-  }
-}
-
-// A clean run (every phase completed or skipped) reads as completed; anything
-// else — a failure or an interruption — reads as failed on the finish screen.
-function overallStatus(metadata: RunMetadata): "completed" | "failed" {
-  const statuses = Object.values(metadata.phases).map((phase) => phase.status)
-  const allDone = statuses.length > 0 && statuses.every((status) => status === "completed" || status === "skipped")
-  return allDone ? "completed" : "failed"
-}
-
-export const __testing = {
-  overallStatus,
-  reconcileAdvisorJournal,
-  replayHistory,
-  snapshotOf,
 }
