+++ src/cli.ts
@@ -961,12 +961,3 @@ ${builtInAgents
   .join("\n")}
 `
 }
-
-export const __testing = {
-  listValue,
-  parseInitArgs,
-  resolveBaseRef,
-  resolveWorktreeOption,
-  splitFlag,
-  writeUpdateResult,
-}
