+++ src/runner.ts
@@ -1584,7 +1584,7 @@ type SessionResult = {
 }
 
 /** Sentinel the closing checkpoint asks a read-only phase to reply with when the advice changes nothing. */
-export const noChangesReply = "NO CHANGES"
+const noChangesReply = "NO CHANGES"
 
 /**
  * Consults the advisor once the phase believes it is finished, and gives it one
@@ -2534,29 +2534,4 @@ function formatSdkError(error: unknown): string {
   }
   if (typeof error === "object" && error && "name" in error) return String((error as { name?: unknown }).name)
   return String(error)
-}
-
-export const __testing = {
-  activity,
-  awaitActionOrAbort,
-  completionFollowUp,
-  describeSessionStatus,
-  describeToolCall,
-  diffSummaryFromEvent,
-  formatCharCount,
-  formatCost,
-  formatModelFromEvent,
-  gateAllowedActions,
-  keepCompletedPhase,
-  payloadID,
-  payloadType,
-  pickString,
-  pulse,
-  rawString,
-  recoveryReport,
-  rememberMessagePartChannel,
-  sleep,
-  stepUsageFromEvent,
-  todosFromEvent,
-  usageFromRecord,
 }
