+++ test/cli.test.ts
@@ -1,24 +1,9 @@
 import { describe, expect, test } from "bun:test"
 import { beforeEach } from "bun:test"
-import { spyOn } from "bun:test"
 
-import { __testing, parseArgs, parseCommand, resolveRunOptions } from "../src/cli"
-
-const { listValue, parseInitArgs, resolveBaseRef, resolveWorktreeOption, splitFlag, writeUpdateResult } = __testing
+import { parseArgs, parseCommand, resolveRunOptions } from "../src/cli"
 
 const validRunID = "20240101-120000-abcd"
-
-const release = {
-  tagName: "v2.0.0",
-  version: { major: 2, minor: 0, patch: 0, prerelease: [] },
-  publishedAt: "2026-01-01T00:00:00Z",
-  assets: [],
-}
-
-const assets = {
-  binary: { name: "convoy-darwin-arm64", browserDownloadUrl: "https://github.com/Inakitajes/convoy/releases/download/v2.0.0/convoy-darwin-arm64" },
-  checksumFile: { name: "SHA256SUMS", browserDownloadUrl: "https://github.com/Inakitajes/convoy/releases/download/v2.0.0/SHA256SUMS" },
-}
 
 describe("parseArgs", () => {
   test("parses a positional prompt", () => {
@@ -266,122 +251,6 @@ describe("parseArgs", () => {
   })
 })
 
-describe("splitFlag", () => {
-  test("splits --flag=value", () => {
-    const { flag, value } = splitFlag("--model=gpt-4")
-    expect(flag).toBe("--model")
-    expect(value).toBe("gpt-4")
-  })
-
-  test("splits --flag with no value", () => {
-    const { flag, value } = splitFlag("--model")
-    expect(flag).toBe("--model")
-    expect(value).toBeUndefined()
-  })
-
-  test("splits short flag with =", () => {
-    const { flag, value } = splitFlag("-f=test.txt")
-    expect(flag).toBe("-f")
-    expect(value).toBe("test.txt")
-  })
-
-  test("handles empty value after =", () => {
-    const { flag, value } = splitFlag("--flag=")
-    expect(flag).toBe("--flag")
-    expect(value).toBe("")
-  })
-
-  test("handles multiple = signs", () => {
-    const { flag, value } = splitFlag("--model=provider/model=extra")
-    expect(flag).toBe("--model")
-    expect(value).toBe("provider/model=extra")
-  })
-})
-
-describe("listValue", () => {
-  test("splits comma-separated values", () => {
-    expect(listValue("a,b,c")).toEqual(["a", "b", "c"])
-  })
-
-  test("handles single value", () => {
-    expect(listValue("a")).toEqual(["a"])
-  })
-
-  test("trims whitespace around items", () => {
-    expect(listValue(" a , b , c ")).toEqual(["a", "b", "c"])
-  })
-
-  test("filters out empty items", () => {
-    expect(listValue("a,,b")).toEqual(["a", "b"])
-  })
-
-  test("returns empty for empty string", () => {
-    expect(listValue("")).toEqual([])
-  })
-
-  test("returns empty for only commas", () => {
-    expect(listValue(",,,")).toEqual([])
-  })
-})
-
-describe("writeUpdateResult", () => {
-  test("source-install status prints message", () => {
-    const writes: string[] = []
-    const spy = spyOn(process.stdout, "write").mockImplementation((chunk: string) => {
-      writes.push(chunk)
-      return true
-    })
-    try {
-      writeUpdateResult({ status: "source-install", message: "source install message" })
-      expect(writes).toEqual(["source install message\n"])
-    } finally {
-      spy.mockRestore()
-    }
-  })
-
-  test("up-to-date status prints version info", () => {
-    const writes: string[] = []
-    const spy = spyOn(process.stdout, "write").mockImplementation((chunk: string) => {
-      writes.push(chunk)
-      return true
-    })
-    try {
-      writeUpdateResult({ status: "up-to-date", currentVersion: "1.0.0", latestVersion: "1.0.0", release, assets })
-      expect(writes).toEqual(["convoy 1.0.0 is up to date (latest: v1.0.0)\n"])
-    } finally {
-      spy.mockRestore()
-    }
-  })
-
-  test("update-available status prints asset info", () => {
-    const writes: string[] = []
-    const spy = spyOn(process.stdout, "write").mockImplementation((chunk: string) => {
-      writes.push(chunk)
-      return true
-    })
-    try {
-      writeUpdateResult({ status: "update-available", currentVersion: "1.0.0", latestVersion: "2.0.0", release, assets })
-      expect(writes).toEqual(["update available: 1.0.0 → v2.0.0 (convoy-darwin-arm64)\n"])
-    } finally {
-      spy.mockRestore()
-    }
-  })
-
-  test("updated status prints success message", () => {
-    const writes: string[] = []
-    const spy = spyOn(process.stdout, "write").mockImplementation((chunk: string) => {
-      writes.push(chunk)
-      return true
-    })
-    try {
-      writeUpdateResult({ status: "updated", currentVersion: "1.0.0", latestVersion: "2.0.0", assetName: "convoy-darwin-arm64" })
-      expect(writes).toEqual(["updated convoy 1.0.0 → v2.0.0 (convoy-darwin-arm64)\n"])
-    } finally {
-      spy.mockRestore()
-    }
-  })
-})
-
 describe("parseCommand", () => {
   test("--help returns help text", async () => {
     const cmd = await parseCommand(["--help"])
@@ -577,101 +446,6 @@ describe("parseCommand", () => {
   test("parses a run command with prompt", async () => {
     const cmd = await parseCommand(["add login"])
     expect(cmd.type).toBe("run")
-  })
-})
-
-describe("parseInitArgs", () => {
-  test("defaults to cwd and non-global", () => {
-    const result = parseInitArgs([])
-    expect(result.global).toBe(false)
-    expect(result.force).toBe(false)
-    expect(result.quiet).toBe(false)
-    expect(result.help).toBeUndefined()
-  })
-
-  test("parses --help", () => {
-    const result = parseInitArgs(["--help"])
-    expect(result.help).toBe(true)
-  })
-
-  test("parses -h", () => {
-    const result = parseInitArgs(["-h"])
-    expect(result.help).toBe(true)
-  })
-
-  test("parses --global", () => {
-    const result = parseInitArgs(["--global"])
-    expect(result.global).toBe(true)
-  })
-
-  test("parses --force", () => {
-    const result = parseInitArgs(["--force"])
-    expect(result.force).toBe(true)
-  })
-
-  test("parses --quiet", () => {
-    const result = parseInitArgs(["--quiet"])
-    expect(result.quiet).toBe(true)
-  })
-
-  test("parses --dir with relative path", () => {
-    const result = parseInitArgs(["--dir", "some/repo"])
-    expect(result.targetDir).toContain("some/repo")
-  })
-
-  test("throws for --global and --dir combo", () => {
-    expect(() => parseInitArgs(["--global", "--dir", "/some/repo"])).toThrow("use either --global or --dir, not both")
-  })
-
-  test("throws for non-flag arg", () => {
-    expect(() => parseInitArgs(["positional"])).toThrow("usage: convoy init")
-  })
-
-  test("throws for unknown flag", () => {
-    expect(() => parseInitArgs(["--unknown"])).toThrow("unknown init flag")
-  })
-
-  test("throws when flag without value is last arg", () => {
-    expect(() => parseInitArgs(["--dir"])).toThrow("requires a value")
-  })
-
-  test("help short-circuits before validation", () => {
-    const result = parseInitArgs(["--global", "--help"])
-    expect(result.help).toBe(true)
-  })
-})
-
-describe("resolveBaseRef", () => {
-  test("uses explicit flag over config defaults", async () => {
-    const parsed = parseArgs(["--base", "develop"])
-    const ref = await resolveBaseRef(parsed, {})
-    expect(ref).toBe("develop")
-  })
-
-  test("uses config default baseRef when no flag", async () => {
-    const parsed = parseArgs([])
-    const ref = await resolveBaseRef(parsed, { baseRef: "main" })
-    expect(ref).toBe("main")
-  })
-})
-
-describe("resolveWorktreeOption", () => {
-  test("uses explicit --worktree flag", async () => {
-    const parsed = parseArgs(["--worktree"])
-    const result = await resolveWorktreeOption(parsed, {})
-    expect(result).toBe(true)
-  })
-
-  test("uses explicit --no-worktree flag", async () => {
-    const parsed = parseArgs(["--no-worktree"])
-    const result = await resolveWorktreeOption(parsed, {})
-    expect(result).toBe(false)
-  })
-
-  test("uses config defaults.worktree", async () => {
-    const parsed = parseArgs([])
-    const result = await resolveWorktreeOption(parsed, { worktree: false })
-    expect(result).toBe(false)
   })
 })
